Add experiment package with a velocity tracker node
The velocity tracker node implements a simple PID that can track only
the linear velocity in the robot frame. The velocity is set using a
service call and send drive_cmd commands. It is designed to use with a
mux to select which output should be used. PID is tunned in simulation and
exposes services to change the gain parameters live.