Add odometry transformer node
This nodes transforms the Pose and Twist information contained in the
odometry message. The odometry message is usually published with reference
to a child_frame_id which is usually the imu frame. This nodes
transforms the pose to a target frame (often base_link) and also
converts velocities to the same target frame. Such transformation is
usefull for the observations that requires the local frame.