Add maximum goal range and handle shutdown
Improved how a couple nodes handle shutdown signals.

Removed the marker that is published by the goal. The goal itself
(PoseStamped) can be used as the main marker in Rviz.

Add a parameter with a maximum range for the goal_publisher node. This
is needed by the waypoint mission to prevent the robot from goin too far
awar from the main track.