Integrate experiment_runner and simulated perception
The experiment runner was publishing only goals (PoseStamped) directly
to the rl_agent. This goals are now using the mux and the experiment
runner also publishes an StampedPoint at the same time. The point can be
used by the simualted perception in the pimage utils to add perception
noise and delay accordingly

The experiment runner also triggers the velocity_controller to reach an
initial velocity defined in the experiment.yaml file. This functionality
was modified to make sure that the target velocity is stable for a given
period. This is important due to the strong inertia of the boat combined
with the thurster delay variation.