Fix issue with missing goals.
It was observed that the goal_mux can drop the first couple goal
messages because the topic is only adverstised once the messages are
arriving.
The experiment runner was changes to publish a couple of goal messages
in the beggining of the experiment to guarantee the mux already
advertises the topic before the actual experiments are running.